Ran tests for: Update NPO text.
Per decision of the PGCA board. This removes a deadlock condition where an NPO that is
recognised could use the name, but an NPO that's being setup cannot yet be recognised.

All NPOs that use the marks in their name should apply for a licence now, as we can approve
that on the condition that they attain and retain recognised status. It also helps strengthen
the marks for us and the NPO.