Ran tests for: Rename Contributor to Significant Contributor
The term "Contributor" was too generic and could easily be confused with
the contributors that are listed in the release notes. Also, we want to
make room for listing more contributors on the website, outside of the
"Recognized Contributors" list.

Also do some copy-editing of the contributor pages:

The Recognized Contributors policy page title gets "Policy" appended
because the old title looked like it was the actual contributors list.

The Contributors Profiles page title gets a "PostgreSQL" prepended so
the "PostgreSQL Contributor" term appears there. (Only the in-page title
is updated since the html header title gets "PostgreSQL:" prepended
automatically anyway.) The section on how to submit additions is moved
to the bottom of the page so people can start looking at the list without
that distraction.

All the sections on that page get short introductory text. (The changes
to data.json will need to be manually transferred to the Django admin
interface.)

Author: Christoph Berg <myon@debian.org>
Reviewed-By: Jonathan S. Katz <jkatz@postgresql.org>